isolated legifrance_to_am
